refactor: Rename ARISA_DSML to creditrisk and reorganize into submodules
- Rename main package `ARISA_DSML` to `creditrisk`.
- Reorganize modules into sub-packages: `core`, `data`, `models`, `utils`.
- Add `__init__.py` files for new sub-packages.
- Move corresponding Python files (`config`, `metrics`, `validation`, `preproc`, `train`, `predict`, `resolve`, `helpers`) into appropriate new sub-packages.
- Update imports in test files (`tests/`) to reflect the new structure.
- Update project structure diagram and usage examples in `README.md`.
- Remove obsolete `credit_default_adaptation_plan.md`.
- Apply code style improvements across modules and notebooks:
    - Use modern type hints (`dict`, `tuple`, `| None`, `-> None`).
    - Replace `exit()` with `sys.exit()`.
    - Use bare `raise` instead of `raise(e)`.
    - Avoid `inplace=True` in notebooks.
    - Minor formatting and cleanup.
