--- ARISA_DSML/config.py
@@ -1,36 +1,0 @@
-"""Config file for module."""
-from pathlib import Path
-
-from dotenv import load_dotenv
-from loguru import logger
-
-# Load environment variables from .env file if it exists
-load_dotenv()
-
-# Paths
-PROJ_ROOT = Path(__file__).resolve().parents[1]
-logger.info(f"PROJ_ROOT path is: {PROJ_ROOT}")
-
-DATASET = "uciml/default-of-credit-card-clients-dataset"
-
-DATA_DIR = PROJ_ROOT / "data"
-RAW_DATA_DIR = DATA_DIR / "raw"
-INTERIM_DATA_DIR = DATA_DIR / "interim"
-PROCESSED_DATA_DIR = DATA_DIR / "processed"
-EXTERNAL_DATA_DIR = DATA_DIR / "external"
-
-MODELS_DIR = PROJ_ROOT / "models"
-
-REPORTS_DIR = PROJ_ROOT / "reports"
-FIGURES_DIR = REPORTS_DIR / "figures"
-
-MODEL_NAME = "credit-default-classifier"
-
-categorical = [
-    "SEX",
-    "EDUCATION",
-    "MARRIAGE",
-    "PAY_0", "PAY_2", "PAY_3", "PAY_4", "PAY_5", "PAY_6"
-]
-
-target = "default.payment.next.month"
