+++ main.py
@@ -1,8 +1,7 @@
-def main():
-    print("Hello from arisa-mlops!")
+def main() -> None:
+    pass
 
 
 if __name__ == "__main__":
     variable = "Hello from arisa-mlops!"
-    print(variable)
     main()
